requirements.txt
#albumentations==1.4.14
albumentations==2.0.8
#datasets==2.18.0
datasets==3.6.0
gradio==5.38.2
ImageHash==4.3.2
matplotlib==3.10.3
numpy==1.21.5
#opencv_contrib_python==4.6.0.66
opencv_contrib_python==4.9.0.80
opencv_contrib_python_headless==4.9.0.80
#opencv_python==4.6.0.66
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.0.3
#Pillow==9.0.1
Pillow==11.3.0
scikit_learn==1.4.0
simple_image_download==0.2
torch==2.3.1
#torch==2.1.2
#torch==1.11.0
torchvision==0.18.1
#torchvision==0.16.2
#torchvision==0.12.0
warmup_scheduler==0.3
